Abstract idx file parsing
IdxFile is the interface to collect all the information that the user needs
to define within an idxfile. a StandardIdxFile implementation is provided
to support the current idx file format. In order to enable a new idx file format
we'll need a dedicated implementation of this interface(we could even ship more
implementations to provide backwards compatiblity).