Simplify input parsing within Command
The Task object now contains all the parameters required by a task. We now have
a different implementation for UserDefinedParameters and for automatically injected
parameters (such as idephix itself). This allow us to create a correct input definition
for the command even if the sign of the task contains parameters that will be injected
at runtime (eg the Idephix instance).

To determine which parameter should receive an IdephixInterface instance we now use
reflection instead of parameter name.