Split IdephixInterface
This splits the generic IdephixInterface into two separate more
specific interfaces. One will be dedicated to executing tasks
and one will be dedicated to building idephix itself (adding tasks,
registering extensions etc etc). This will ease the definition of
a context that we can pass to the tasks instead of the whole Idephix
instance